feat: Introduce support MongoDB CSFLE
Introduce support for MongoDB Client-Side Field Level Encryption (CSFLE).

CSFLE provides enhanced security for data stored in MongoDB by allowing field-level encryption. The configuration for CSFLE is now made easy with auto-encryption via the `.karenc.yml` file, which specifies the fields to be encrypted.

Additionally, the instantiation of the MongoDB Client has been moved from the constructor to the `connect` method. This change is necessary because encryption requires authentication information (e.g. DB, KMS, etc.) to build the schemaMap for encryption, which may not be available during object instantiation.